chore(storybook): remove next folders move files inline
