+++ packages/react/src/components/OverflowMenuV2/OverflowMenuv2.stories.js
@@ -8,10 +8,10 @@
 import React from 'react';
 import { ArrowsVertical } from '@carbon/icons-react';
 
-import Menu from '../../Menu';
-import { StoryFrame, buildMenu } from '../../Menu/_storybook-utils';
+import Menu from '../Menu';
+import { StoryFrame, buildMenu } from '../Menu/_storybook-utils';
 
-import { OverflowMenuV2 } from '..';
+import { OverflowMenuV2 } from '.';
 
 export default {
   title: 'Experimental/unstable_Menu/OverflowMenuV2',
